requirements.txt
cryptography
requests
python-dotenv
pandas
numpy
websockets

